Adding user logout functionality. Re-organizing components.
