+++ client/package.json
@@ -23,6 +23,7 @@
     "pdfjs-dist": "^5.4.149",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
+    "react-router-dom": "^7.9.1",
     "typescript": "^5.9.2"
   },
   "devDependencies": {
